[RENAME] BaseEntity 디렉토리 위치 이동
