+++ src/main/java/kr/hhplus/be/server/domain/product/Product.java
@@ -1,7 +1,7 @@
 package kr.hhplus.be.server.domain.product;
 
 import jakarta.persistence.*;
-import kr.hhplus.be.server.common.BaseTimeEntity;
+import kr.hhplus.be.server.common.config.BaseTimeEntity;
 import kr.hhplus.be.server.common.exception.OutOfStockException;
 import lombok.*;
